implementing the congestion model for group, region and data center

+++ Car.hpp
@@ -17,8 +17,8 @@ class Car
 {
   public:
     int car_spot_number;
-    int car_cluster_number;
-    int car_group_number, car_region_number;
+    signed short int car_cluster_number;
+    signed short int car_group_number, car_region_number;
     int arrival_time_of_car;
     int departure_time_of_car;
     int job_number;
